:recycle: [#46] refactor: Post의 Service 계층이 다른 도메인의 영속성 계층에 의존하는 문제점 리팩토링
