+++ src/main/java/gaji/service/domain/common/service/HashtagService.java
@@ -1,13 +1,17 @@
 package gaji.service.domain.common.service;
 
 
+import gaji.service.domain.common.entity.Hashtag;
 import gaji.service.domain.common.entity.SelectHashtag;
 import gaji.service.domain.enums.PostTypeEnum;
 
 import java.util.List;
 
-public interface HashtagQueryService {
+public interface HashtagService {
 
     boolean ExistHashtagByName(String name);
     List<SelectHashtag> findAllFetchJoinWithCategoryByEntityIdAndPostType(Long entityId, PostTypeEnum postType);
+    List<Hashtag> createHashtagEntityList(List<String> hashtagStringList);
+    void saveAllSelectHashtag(List<SelectHashtag> selectHashtagList);
+    void deleteAllByEntityIdAndType(Long entityId, PostTypeEnum postType);
 }
